requirements.txt
fastapi
uvicorn
pyjwt
bcrypt 
sqlalchemy
python-dotenv
pytest
httpx
aiosqlite
pytest-asyncio